+++ src/contents.rs
@@ -372,6 +372,12 @@ impl AsMut<[u8]> for Contents {
     }
 }
 
+impl AsMut<ContentsRef> for Contents {
+    fn as_mut(&mut self) -> &mut ContentsRef {
+        self
+    }
+}
+
 impl Borrow<ContentsRef> for Contents {
     fn borrow(&self) -> &ContentsRef {
         self