+++ CHANGELOG.md
@@ -131,6 +131,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
     - 'PartialOrd<T>' for 'IdRef' where 'T' is bounded on 'Borrow<IdRef>'
     - 'PartialOrd' for Length
     - 'AsRef<ContentsRef>' for Contents
+    - 'AsMut<ContentsRef>' for Contents
     - 'Index<T>' for ContentsRef where 'T' is bounded on 'SliceIndex<[u8]>'
 - The following functions and methods
     - Ber::new\_indefinite()